python/requirements.txt
PyYAML==3.10
argparse==1.2.1
pyelftools==0.21
wsgiref==0.1.2
